+++ modules/logging/core/src/sap/commerce/toolset/logging/CxLogService.kt
@@ -96,6 +96,27 @@ class CxLogService(private val project: Project, private val coroutineScope: Cor
         }
     }
 
+    fun deleteLogger(templateUUID: String, loggerName: String) = with(CxCustomLogTemplatesSettings.getInstance(project)) {
+        val loggerTemplateState = templates
+            .find { it.uuid == templateUUID }
+            ?.mutable()
+            ?.apply {
+                //remove logger
+                val newLoggerConfigs = loggers.get().filter { logger -> logger.name.get() != loggerName }
+
+                loggers.set(newLoggerConfigs)
+            }
+            ?.immutable()
+            ?: return@with
+
+        updateCustomLoggerTemplateInternal(loggerTemplateState)
+
+        coroutineScope.launch {
+            val modifiedTemplate = loggerTemplateState.presentation(project)
+            project.messageBus.syncPublisher(CxCustomLogTemplateStateListener.TOPIC).onLoggerUpdated(modifiedTemplate)
+        }
+    }
+
     private fun updateCustomLoggerTemplateInternal(template: CxCustomLogTemplateState) {
         with(CxCustomLogTemplatesSettings.getInstance(project)) {
             templates = templates.toMutableList().apply {